requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
anthropic==0.27.0
chromadb==0.4.24
sentence-transformers==2.2.2
python-dotenv==1.0.0
jinja2==3.1.2
pyyaml==6.0.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.0
freezegun==1.4.0
